+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -148,7 +148,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-11T17:31:13.599Z",
+    "2026-02-12T05:19:10.507Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {